+++ src/core/lang/sv/primaryunit.rs
@@ -16,8 +16,10 @@
 //
 
 use super::{symbols::SystemVerilogSymbol, token::identifier::Identifier};
+use crate::core::lang::vhdl::primaryunit::HdlNamingError;
 use crate::{core::lang::sv::symbols::SystemVerilogParser, util::anyerror::CodeFault};
 use std::collections::HashMap;
+use std::path::PathBuf;
 use std::str::FromStr;
 
 #[derive(PartialEq, Hash, Eq, Debug)]
@@ -131,6 +133,7 @@ pub fn collect_units(files: &Vec<String>) -> Result<HashMap<Identifier, PrimaryU
     for source_file in files {
         // only read the HDL files
         if crate::core::fileset::is_systemverilog(&source_file) == true {
+            // println!("parsing sv file: {}", source_file);
             // parse text into Verilog symbols
             let contents = match std::fs::read_to_string(&source_file) {
                 Ok(dump) => dump,
@@ -207,8 +210,28 @@ pub fn collect_units(files: &Vec<String>) -> Result<HashMap<Identifier, PrimaryU
                 })
                 .collect();
 
-            // push to the global list
-            result.extend(units);
+            // push to the global list (ensure there are zero duplicate names)
+            for (_key, primary) in units {
+                if let Some(dupe) = result.insert(primary.get_name().clone(), primary) {
+                    return Err(CodeFault(
+                        None,
+                        Box::new(HdlNamingError::DuplicateIdentifier(
+                            dupe.get_name().to_string(),
+                            PathBuf::from(source_file),
+                            result
+                                .get(dupe.get_name())
+                                .unwrap()
+                                .get_unit()
+                                .get_symbol()
+                                .unwrap()
+                                .get_position()
+                                .clone(),
+                            PathBuf::from(dupe.get_unit().get_source_file()),
+                            dupe.get_unit().get_symbol().unwrap().get_position().clone(),
+                        )),
+                    ))?;
+                }
+            }
         }
     }
     Ok(result)